Feature/bun nginx (#217)
* feat: added bun

* machineID added for product (#214)

* machineID added for product

* additives added, and tested

---------


* feat: dto agnostic sanitizing (#213)

* feature/soft delete restrictions

* feat: kiosk updates

* feat: added region manager view

* feature/restrict soft-deletions

* feat: added region and franchisee shared views

* feat: added welcome message, fixed number

* feat: loaders

* feat: added admin list loader

* fix/region context filters

* updated logger (#201)

* on delete store additive check for store product size default additive (#197)

* on delete store additive check for store product size default additive

* removed redundant newline

* passed manual tests and removed unnecessary comments

* improved according to comments

* added local error

* handled soft deletes

---------


* updated logger

* hardcoded log level

* hotfix

* fixed golangci errcheck

* temporar solution for ci

* temporar solution for ci at stockRequest

* temporar solution for ci at order

---------


* Feature/restrict deletions (#200)

* feature/soft delete restrictions

* feature/restrict soft-deletions

* fix/region context filters

---------


* feature/store context filters

* proper messaging: categories, warehouses, warehouse-stock, suppliers (#202)

* on delete store additive check for store product size default additive (#197)

* on delete store additive check for store product size default additive

* removed redundant newline

* passed manual tests and removed unnecessary comments

* improved according to comments

* added local error

* handled soft deletes

---------


* feat: categories, warehouses, warehouse-stock, suppliers

* feat: stores messages

* feat: regions

* fix: abandon duplicates

---------


* feat: added loaders on lists (#205)

* Feature/restrict deletions (#204)

* feature/soft delete restrictions

* feature/restrict soft-deletions

* fix/region context filters

* feature/store context filters

---------


* fix/test & float values for product size and additives

* Feature/restrict deletions (#206)

* feature/soft delete restrictions

* feature/restrict soft-deletions

* fix/region context filters

* feature/store context filters

* fix/test & float values for product size and additives

---------


* feat: added list loader toolbar on top (#207)

* fix/warehouse address update

* fix/assigning

* fix/restrictions

* feat: dto agnostic sanitizing

* Update export.orders_v2.go

* fix/tests

* Revert "Merge branch 'dev' into feature/restrict-deletions"

This reverts commit dfd052bf6b2e0253503d4a30526f23238fc05961, reversing
changes made to 4b64f7cd03d32325f926efcd568f7237a169d3ee.

* fix/tests

* fixes

* fix linter

* Revert "Merge remote-tracking branch 'origin/feature/restrict-deletions' into hotfix/sanitizer"

This reverts commit 46a692d152d1bc190eb95bb89aeeed3e06a3daee, reversing
changes made to 0efe35ab86deed2b24627086c7f1c2d9844eeb72.

---------


* Refactor/tests (#216)

* feature/soft delete restrictions

* feat: kiosk updates

* feat: added region manager view

* feature/restrict soft-deletions

* feat: added region and franchisee shared views

* feat: added welcome message, fixed number

* feat: loaders

* feat: added admin list loader

* fix/region context filters

* updated logger (#201)

* on delete store additive check for store product size default additive (#197)

* on delete store additive check for store product size default additive

* removed redundant newline

* passed manual tests and removed unnecessary comments

* improved according to comments

* added local error

* handled soft deletes

---------


* updated logger

* hardcoded log level

* hotfix

* fixed golangci errcheck

* temporar solution for ci

* temporar solution for ci at stockRequest

* temporar solution for ci at order

---------


* Feature/restrict deletions (#200)

* feature/soft delete restrictions

* feature/restrict soft-deletions

* fix/region context filters

---------


* feature/store context filters

* proper messaging: categories, warehouses, warehouse-stock, suppliers (#202)

* on delete store additive check for store product size default additive (#197)

* on delete store additive check for store product size default additive

* removed redundant newline

* passed manual tests and removed unnecessary comments

* improved according to comments

* added local error

* handled soft deletes

---------


* feat: categories, warehouses, warehouse-stock, suppliers

* feat: stores messages

* feat: regions

* fix: abandon duplicates

---------


* feat: added loaders on lists (#205)

* Feature/restrict deletions (#204)

* feature/soft delete restrictions

* feature/restrict soft-deletions

* fix/region context filters

* feature/store context filters

---------


* fix/test & float values for product size and additives

* Feature/restrict deletions (#206)

* feature/soft delete restrictions

* feature/restrict soft-deletions

* fix/region context filters

* feature/store context filters

* fix/test & float values for product size and additives

---------


* feat: added list loader toolbar on top (#207)

* fix/warehouse address update

* fix/assigning

* fix/restrictions

* fix/tests

* Revert "Merge branch 'dev' into feature/restrict-deletions"

This reverts commit dfd052bf6b2e0253503d4a30526f23238fc05961, reversing
changes made to 4b64f7cd03d32325f926efcd568f7237a169d3ee.

* fix/tests

* fixes

* fix linter

* refactor

* refactor

---------


* fix: fixed some Vite errors related to render

* feat: added machine id, qr generator, bun build

---------

+++ backend/internal/modules/warehouse/stockMaterial/stockMaterialCategory/stockMaterialCategory.handler.go
@@ -137,7 +137,7 @@ func (h *StockMaterialCategoryHandler) Delete(c *gin.Context) {
 		return
 	}
 
-	response, err := h.service.GetByID(uint(id))
+	response, err := h.service.GetByID(id)
 	if err != nil {
 		if errors.Is(err, types.ErrStockMaterialCategoryNotFound) {
 			localization.SendLocalizedResponseWithKey(c, types.Response404StockMaterialCategory)
@@ -146,7 +146,7 @@ func (h *StockMaterialCategoryHandler) Delete(c *gin.Context) {
 		return
 	}
 
-	if err := h.service.Delete(uint(id)); err != nil {
+	if err := h.service.Delete(id); err != nil {
 		if errors.Is(err, types.ErrStockMaterialCategoryIsInUse) {
 			localization.SendLocalizedResponseWithKey(c, types.Response409StockMaterialCategoryDeleteInUse)
 			return
@@ -157,7 +157,7 @@ func (h *StockMaterialCategoryHandler) Delete(c *gin.Context) {
 
 	action := types.DeleteStockMaterialAuditFactory(
 		&data.BaseDetails{
-			ID:   uint(id),
+			ID:   id,
 			Name: response.Name,
 		})
